# Performance: Drag items via transform property to avoid layout and re-paints