--- a/packages/components/src/draggable/index.js
+++ b/packages/components/src/draggable/index.js
@@ -3,6 +3,11 @@
  */
 import { useEffect, useRef } from '@wordpress/element';
 
+/**
+ * External dependencies
+ */
+import { throttle } from 'lodash';
+
 const dragImageClass = 'components-draggable__invisible-drag-image';
 const cloneWrapperClass = 'components-draggable__clone';
 const cloneHeightTransformationBreakpoint = 700;
@@ -18,14 +23,30 @@ const bodyClass = 'is-dragging-components-draggable';
 /**
  * @typedef Props
  * @property {(props: RenderProp) => JSX.Element | null}  children                         Children.
- * @property {(event: import('react').DragEvent) => void} [onDragStart]                    Callback when dragging starts.
- * @property {(event: import('react').DragEvent) => void} [onDragOver]                     Callback when dragging happens over the document.
+ * @property {(event: import('react').DragEvent) => void} [onDragStart]                    Callback when dragging
+ *     starts.
+ * @property {(event: import('react').DragEvent) => void} [onDragOver]                     Callback when dragging
+ *     happens over the document.
  * @property {(event: import('react').DragEvent) => void} [onDragEnd]                      Callback when dragging ends.
- * @property {string}                                     [cloneClassname]                 Classname for the cloned element.
+ * @property {string}                                     [cloneClassname]                 Classname for the cloned
+ *     element.
  * @property {string}                                     [elementId]                      ID for the element.
- * @property {any}                                        [transferData]                   Transfer data for the drag event.
- * @property {string}                                     [__experimentalTransferDataType] The transfer data type to set.
- * @property {import('react').ReactNode}                  __experimentalDragComponent      Component to show when dragging.
+ * @property {any}                                        [transferData]                   Transfer data for the drag
+ *     event.
+ * @property {string}                                     [__experimentalTransferDataType] The transfer data type to
+ *     set.
+ * @property {import('react').ReactNode}                  __experimentalDragComponent      Component to show when
+ *     dragging.
+ */
+
+/**
+ * @typedef DragState
+ * @property {number}                    x            Current X coordinate of cloneWrapper
+ * @property {number}                    y            Current Y coordinate of cloneWrapper
+ * @property {number}                    cursorLeft   Last mouse x coordinate
+ * @property {number}                    cursorTop    Last mouse y coordinate
+ * @property {number}                    scale        Scale to apply to the draggable.
+ * @property {HTMLDivElement}            cloneWrapper Cloned Draggable
  */
 
 /**
@@ -62,6 +83,29 @@ export default function Draggable( {
 	}
 
 	/**
+	 * @param {import('react').DragEvent} e
+	 * @param {DragState} dragState
+	 */
+	function over( e, dragState ) {
+		const nextX = dragState.x + e.clientX - dragState.cursorLeft;
+		const nextY = dragState.y + e.clientY - dragState.cursorTop;
+		if ( dragState.scale === 1 ) {
+			dragState.cloneWrapper.style.transform = `translate( ${ nextX }px, ${ nextY }px )`;
+		} else {
+			dragState.cloneWrapper.style.transform = `translate( ${ nextX }px, ${ nextY }px ) scale( ${ dragState.scale } )`;
+		}
+
+		// Update cursor coordinates.
+		dragState.cursorLeft = e.clientX;
+		dragState.cursorTop = e.clientY;
+		dragState.x = nextX;
+		dragState.y = nextY;
+		if ( onDragOver ) {
+			onDragOver( e );
+		}
+	}
+
+	/**
 	 * This method does a couple of things:
 	 *
 	 * - Clones the current element and spawns clone over original element.
@@ -81,6 +125,10 @@ export default function Draggable( {
 		);
 
 		const cloneWrapper = ownerDocument.createElement( 'div' );
+		// Reset position to 0,0. Natural stacking order will position this lower, even with a transform otherwise.
+		cloneWrapper.style.top = 0;
+		cloneWrapper.style.left = 0;
+
 		const dragImage = ownerDocument.createElement( 'div' );
 
 		// Set a fake drag image to avoid browser defaults. Remove from DOM
@@ -98,12 +146,16 @@ export default function Draggable( {
 			cloneWrapper.classList.add( cloneClassname );
 		}
 
+		let x,
+			y,
+			scale = 1;
 		// If a dragComponent is defined, the following logic will clone the
 		// HTML node and inject it into the cloneWrapper.
 		if ( dragComponentRef.current ) {
 			// Position dragComponent at the same position as the cursor.
-			cloneWrapper.style.top = `${ event.clientY }px`;
-			cloneWrapper.style.left = `${ event.clientX }px`;
+			x = event.clientX;
+			y = event.clientY;
+			cloneWrapper.style.transform = `translate( ${ x }px, ${ y }px )`;
 
 			const clonedDragComponent = ownerDocument.createElement( 'div' );
 			clonedDragComponent.innerHTML = dragComponentRef.current.innerHTML;
@@ -129,19 +181,19 @@ export default function Draggable( {
 
 			if ( elementRect.height > cloneHeightTransformationBreakpoint ) {
 				// Scale down clone if original element is larger than 700px.
-				cloneWrapper.style.transform = 'scale(0.5)';
 				cloneWrapper.style.transformOrigin = 'top left';
 				// Position clone near the cursor.
-				cloneWrapper.style.top = `${ event.clientY - 100 }px`;
-				cloneWrapper.style.left = `${ event.clientX }px`;
+				x = event.clientX;
+				y = event.clientY - 100;
+				// Can also be written as the affine transform: matrix( 0.5, 0, 0, 0.5, x, y ).
+				// Note that order matters, translate should be done before scale.
+				cloneWrapper.style.transform = `translate( ${ x }px, ${ y }px ) scale(0.5)`;
+				scale = 0.5;
 			} else {
 				// Position clone right over the original element (20px padding).
-				cloneWrapper.style.top = `${
-					elementTopOffset - clonePadding
-				}px`;
-				cloneWrapper.style.left = `${
-					elementLeftOffset - clonePadding
-				}px`;
+				x = elementLeftOffset - clonePadding;
+				y = elementTopOffset - clonePadding;
+				cloneWrapper.style.transform = `translate( ${ x }px, ${ y }px )`;
 			}
 
 			// Hack: Remove iFrames as it's causing the embeds drag clone to freeze
@@ -156,30 +208,37 @@ export default function Draggable( {
 		}
 
 		// Mark the current cursor coordinates.
-		let cursorLeft = event.clientX;
-		let cursorTop = event.clientY;
+		const dragState = {
+			x,
+			y,
+			cursorLeft: event.clientX,
+			cursorTop: event.clientY,
+			cloneWrapper,
+			scale,
+		};
+
+		/**
+		 * @param {import("react").DragEvent<Element>} e
+		 */
+		function update( e ) {
+			//Skip doing any work if mouse has not moved.
+			if (
+				dragState.cursorLeft === e.clientX &&
+				dragState.cursorTop === e.clientY
+			) {
+				return;
+			}
+			over( e, dragState );
+		}
+
+		// Aim for 60fps (16 ms per frame) for now. We can potentially use requestAnimationFrame (raf) instead,
+		// note that browsers may throttle raf below 60fps in certain conditions.
+		const throttledDragOver = throttle( update, 16 );
 
 		/**
 		 * @param {import('react').DragEvent} e
 		 */
-		function over( e ) {
-			cloneWrapper.style.top = `${
-				parseInt( cloneWrapper.style.top, 10 ) + e.clientY - cursorTop
-			}px`;
-			cloneWrapper.style.left = `${
-				parseInt( cloneWrapper.style.left, 10 ) + e.clientX - cursorLeft
-			}px`;
-
-			// Update cursor coordinates.
-			cursorLeft = e.clientX;
-			cursorTop = e.clientY;
-
-			if ( onDragOver ) {
-				onDragOver( e );
-			}
-		}
-
-		ownerDocument.addEventListener( 'dragover', over );
+		ownerDocument.addEventListener( 'dragover', throttledDragOver );
 
 		// Update cursor to 'grabbing', document wide.
 		ownerDocument.body.classList.add( bodyClass );
@@ -208,7 +267,7 @@ export default function Draggable( {
 			// Reset cursor.
 			ownerDocument.body.classList.remove( bodyClass );
 
-			ownerDocument.removeEventListener( 'dragover', over );
+			ownerDocument.removeEventListener( 'dragover', throttledDragOver );
 
 			clearTimeout( timerId );
 		};