### Background 

Our draggable component updates a drag chip by modifying `top` and `left` properties. This unfortunately forces the browser to go through layout > paint> and compositing steps which is expensive. As we can see in the profiler on trunk, we're both dropping frames and seeing layout shifts.

Changes in this PR update this to use `transform` instead which can skip layout and paint, [to go straight to composite](https://developers.google.com/web/fundamentals/performance/rendering). We also add a throttle to reduce the number of fired drag events, and early escape if the mouse has not moved.

Our drop zone handlers are also causing major performance issues when dragging quickly. This will be handled in https://github.com/WordPress/gutenberg/pull/33435

Videos below are taken with the Armando theme, using the "alternating rows with images and headings" pattern.

| before | after |
|-----|-----|
| <img width="1350" alt="Screen Shot 2021-07-14 at 2 33 23 PM" src="https://user-images.githubusercontent.com/1270189/125696399-503b95b7-0e33-409e-82ab-6c3a0892a04b.png"> | <img width="1156" alt="Screen Shot 2021-07-14 at 2 26 17 PM" src="https://user-images.githubusercontent.com/1270189/125695528-41e3ecb7-4652-4486-b799-29879189ffc4.png"> |

### Testing Instructions

- Verify that there are no regressions with dragging
- When profiling verify that layout shifts are visible in trunk when dragging, but not visible in this branch. (See examples in video).

#### Profiling In Trunk 

- Notice how the layout shifts are coming from the draggable chip.
- The lionshare of handling time is still taken up by dropzone in some cases. This will be addressed in a follow up.

https://user-images.githubusercontent.com/1270189/125692895-71c7d498-928f-43cf-8db4-a4abbf70558c.mp4

#### Profiling in Branch
- No more layout shifts visible from profiler
- Dragging over multiple dropzones is still slow and will be addressed in a follow up. 

https://user-images.githubusercontent.com/1270189/125694389-c3d99e7d-53ed-498d-912b-0e4aa73c0247.mp4