@gwwar 2021-07-13
Playing around with the draggable + dropzone code, it feels like we have a few opportunities.

- We can add a single listener that fires on document dragover to track if our mouse is moving quickly, or not (if the user is hovering). If the user is not hovering we can avoid trying to trigger all these dropzones from rendering/animating.
- Providing a bigger drag hint + drop zone target would be a huge improvement. Maybe https://github.com/WordPress/gutenberg/pull/33132 would help with this.
- Right now **too many dragover events are firing**. I'll continue to look for spots to throttle.
- Since normally users can only drag one item onto a single droppable zone, it may make sense to convert logic to a global listener. The logic is currently a bit hard to follow (and likely more error prone with multiple child listeners).