@talldan 2021-07-14
I did some testing and the results seem good so far. 

The blue drop zone indicators are noticeably quite delayed at updating position, so that might be something to improve. I think that's one of those things that needs to feel quite responsive, or the user will perceive it as slow, even if the code is well optimized.

Definitely a good idea to work on this iteratively, so if we can get some noticeable improvements with the Draggable first, we can ship that straight away.