@gwwar 2021-07-14
> Definitely a good idea to work on this iteratively, so if we can get some noticeable improvements with the Draggable first, we can ship that straight away.

Sounds good, I can tidy up changes to pull out the draggable changes first. General dragging performance is still going to be pretty slow from the existing dropzone handlers, but the difference in removing those layout shifts should be visible from browser profiling.

I'll spin up a follow up draft to continue experimenting with drag intents.