@gwwar 2021-07-14
This one is ready for review 👍  I moved over other exploratory dropzone performance handling to https://github.com/WordPress/gutenberg/pull/33435